refactor: extract linkPathToPeerVersion into its own file
